+++ precode_test.go
@@ -4,89 +4,58 @@ import (
 	"fmt"
 	"github.com/stretchr/testify/require"
 	"io"
-	"log/slog"
 	"net/http"
 	"net/http/httptest"
-	"os"
-	"strconv"
 	"strings"
 	"testing"
 )
-
-var cafeList = map[string][]string{
-	"moscow": []string{"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
-}
-
-func mainHandle(w http.ResponseWriter, req *http.Request) {
-	countStr := req.URL.Query().Get("count")
-	if countStr == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("count missing"))
-		return
-	}
-
-	count, err := strconv.Atoi(countStr)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong count value"))
-		return
-	}
-
-	city := req.URL.Query().Get("city")
-
-	cafe, ok := cafeList[city]
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong city value"))
-		return
-	}
-
-	if count > len(cafe) {
-		count = len(cafe)
-	}
-
-	answer := strings.Join(cafe[:count], ",")
-
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte(answer))
-}
-
-var logger = slog.New(slog.NewJSONHandler(os.Stdout, nil))
 
 func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
 	totalCount := 4
 	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/cafe?count=%d&city=moscow", 10), nil)
 
-	responseRecorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(mainHandle)
-	handler.ServeHTTP(responseRecorder, req)
-
-	res := responseRecorder.Result()
+	res := getResponse(req)
 	defer res.Body.Close()
 
 	require.Equal(t, http.StatusOK, res.StatusCode)
 
 	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		logger.Error("Couldn't read body from response", "error", err.Error())
-	}
-
-	require.NotEmpty(t, body, "Response body should be not empty")
-	s := string(body)
-	items := strings.Split(s, ",")
-
+	require.NoError(t, err)
+	items := strings.Split(string(body), ",")
 	require.Len(t, items, totalCount, "Returned Coffee items count should be 4")
 }
 
 func TestMainHandlerWhenCityIsNotMoscow(t *testing.T) {
 	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/cafe?count=%d&city=asdf", 10), nil)
 
+	res := getResponse(req)
+	defer res.Body.Close()
+
+	body, err := io.ReadAll(res.Body)
+
+	require.NoError(t, err)
+	require.Equal(t, "wrong count value", string(body))
+	require.Equal(t, http.StatusBadRequest, res.StatusCode)
+}
+
+func TestMainHandlerBodyShouldBeNotEmpty(t *testing.T) {
+	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/cafe?count=%d&city=moscow", 10), nil)
+
+	res := getResponse(req)
+	defer res.Body.Close()
+
+	require.Equal(t, http.StatusOK, res.StatusCode)
+
+	body, err := io.ReadAll(res.Body)
+	require.NoError(t, err)
+	require.NotEmpty(t, body, "Response body should be not empty")
+}
+
+func getResponse(req *http.Request) *http.Response {
 	responseRecorder := httptest.NewRecorder()
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
 
 	res := responseRecorder.Result()
-	defer res.Body.Close()
-
-	require.Equal(t, http.StatusBadRequest, res.StatusCode)
+	return res
 }
